requirements.txt
streamlit>=1.28.0
requests>=2.31.0
gspread>=5.12.0
oauth2client>=4.1.3
openai>=1.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2

